web-audio/audio-destination-node.html

<!doctype html>
<html class="a">
 <head>
  <title>AudioDestinationNode interface</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/lodash.js"></script>
  <script src="/resources/vendor-prefixes.js"></script>
  <script src="/resources/helpers.js"></script>
 </head>
 <body class="a">
  <div id="log"></div>
  <script>

var context;
setup(function() {
  context = new AudioContext();
});

test(function() {
  assert_defined(context, "context is defined")

  var destination = context.destination;
  assert_equals(destination.numberOfInputs, 1, "should have 1 input");
  assert_equals(destination.numberOfOutputs, 0, "should have 0 outputs");
  assert_equals(destination.numberOfChannels, 2, "should have 2 channels by default");
}, "destination");

  </script>
 </body>
</html>
